#1248 Make service provider properties write-only in ApplicationContext (internal get)
